+++ src/lib/trigger-contract.js
@@ -1,0 +1,3 @@
+'use strict';
+
+module.exports = require('../../packages/skillcanary/src/lib/trigger-contract');